store then show paylog

+++ cchack.hh
@@ -24,7 +24,7 @@ $ gcc cchack.c -lnfc -o readnfccc
 
 #define MAX_FRAME_LEN 300
 
-#define HEX(c) std::hex << std::uppercase << std::setw(2) << std::setfill('0') << c << std::dec
+#define HEX(c) std::hex << std::uppercase << std::setw(2) << std::setfill('0') << (unsigned int)c << std::dec
 
 typedef unsigned char byte_t;
